requirements.txt
amqp==2.3.2
asn1crypto==0.24.0
backports.shutil-get-terminal-size==1.0.0
bcrypt==3.1.3
billiard==3.5.0.4
celery==4.0.2
cffi==1.11.5
cryptography==2.1.4
decorator==4.3.0
Django==1.11.13
django-filter==1.0.4
django-rest-framework==0.1.0
django-smtp-ssl==1.0
djangorestframework==3.8.2
enum34==1.1.6
gunicorn==19.7.1
idna==2.7
ipaddress==1.0.22
ipython==5.8.0
ipython-genutils==0.2.0
kombu==4.0.2
Markdown==2.6.9
pathlib2==2.3.2
pexpect==4.6.0
pickleshare==0.7.4
Pillow==2.3.0
prompt-toolkit==1.0.15
psycopg2==2.7.5
ptyprocess==0.6.0
pyasn1==0.4.4
pycparser==2.18
Pygments==2.2.0
PyJWT==1.6.4
pytz==2018.4
redis==2.10.5
scandir==1.8
simplegeneric==0.8.1
six==1.11.0
traitlets==4.3.2
vine==1.1.4
wcwidth==0.1.7
whitenoise==3.3.1
